+++ src/index.html
@@ -86,6 +86,7 @@
 
     <script type="module" src="js/scripts/main/burger_menu/extract_root_container.js"></script>
     <script type="module" src="js/scripts/main/burger_menu/extract_menu_template.js"></script>
+    <script type="module" src="js/scripts/main/burger_menu/fold_menu_function.js"></script>
     <script type="module" src="js/scripts/main/burger_menu/burger_menu.js"></script>
     <script type="module" src="js/scripts/main/burger_menu/hover_animation_burger_menu.js"></script>
