+++ src/sass/main/burger_menu.sass
@@ -21,7 +21,6 @@
     align-items: center
     justify-content: center
     padding-bottom: 10%
-    border: 3px solid black
     &__navigation-container
       margin-top: 8%
       display: flex
